requirements.txt
# Requirements for the Call Tracking sample Django project

# Django (1.8 is a long-term support release)
Django==2.2.1

# Third-party app dependencies
django-bootstrap4==0.0.8
django-icons==0.2.1
django-forms-bootstrap==3.1.0
django-phonenumber-field==2.4.0
phonenumbers==8.10.12
setuptools==41.0.1
psycopg2==2.8.4
twilio==6.27.0
whitenoise==4.1.2

# Production dependencies
dj-database-url==0.5.0
gunicorn==19.9.0

# Test dependencies
coverage==4.5.2
mock==3.0.5
model-mommy==1.6.0
six==1.12.0
